Refactors xmlrunner test out into test directory
